requirements.txt
streamlit==1.29.0
transformers==4.35.0
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.0
numpy==1.24.4
nest_asyncio==1.5.6
protobuf==3.20.*
